requirements.txt
numpy
tqdm
pydantic>=2.0.0
python-dotenv
requests
PyMuPDF
pytesseract
Pillow
faiss-cpu
sentence-transformers
openai



